* Response to code review   * Moved files in graph-utils into utilities and deleted graph-utils   * Used nesting in graph.scss   * Moved graph.test into parent folder
