+++ v3/src/components/graph/components/dotplotdots.tsx
@@ -5,7 +5,7 @@ import {plotProps, InternalizedData, defaultRadius, defaultDiameter, dragRadius}
   from "../graphing-types"
 import {useDragHandlers, useSelection} from "../hooks/graph-hooks"
 import {IDataSet} from "../../../data-model/data-set"
-import {getScreenCoord, setPointCoordinates} from "../graph-utils/graph_utils"
+import {getScreenCoord, setPointCoordinates} from "../utilities/graph_utils"
 
 
 export const DotPlotDots = memo(observer(function DotPlotDots(props: {
