+++ v3/src/components/graph/components/graph.test.tsx
@@ -2,7 +2,7 @@ import { render, screen } from "@testing-library/react"
 import React from "react"
 import { DataBroker } from "../../../data-model/data-broker"
 import { DataSet, toCanonical } from "../../../data-model/data-set"
-import {Graph} from "../components/graph"
+import {Graph} from "./graph"
 
 describe("Graph", () => {
   let broker: DataBroker
