+++ v3/src/components/graph/components/scatterdots.tsx
@@ -3,7 +3,7 @@ import {select} from "d3"
 import {plotProps, InternalizedData, defaultRadius, dragRadius} from "../graphing-types"
 import {useDragHandlers, useSelection} from "../hooks/graph-hooks"
 import {IDataSet} from "../../../data-model/data-set"
-import {getScreenCoord, setPointCoordinates} from "../graph-utils/graph_utils"
+import {getScreenCoord, setPointCoordinates} from "../utilities/graph_utils"
 
 export const ScatterDots = memo(function ScatterDots(props: {
   plotProps: plotProps,
